requirements.txt
requests
tavily-python
openai
python-dotenv
loguru